Rename "reusable" components to "common" components
