+++ src/components/ControlPanel/RegionSelector/index.tsx
@@ -6,7 +6,7 @@ import '../../../style/dropdownForm.css';
 import SuperRegionSelector from './SuperRegionSelector';
 import SubRegionCollectionSelector from './SubRegionCollectionSelector';
 import SubRegionSelector from './SubRegionSelector';
-import LoadingButton from '../../reusable/LoadingButton';
+import LoadingButton from '../../common/LoadingButton';
 import selectedSuperRegionNameAtom from '../../../state/client/selectedSuperRegionName';
 import selectedSuperRegionAtom from '../../../state/client/derived/selectedSuperRegion';
 import useRegionsIndexQuery from '../../../serverState/regionsIndex';
